requirements.txt
azure-ai-projects
azure-identity
requests
fastapi
uvicorn